+++ matrix.h
@@ -12,17 +12,17 @@
 
 class Matrix {
 
-    std::vector<std::vector<uint64_t>> Column_Index;
+    std::vector<std::vector<size_t>> Column_Index;
     std::vector<std::vector<double>> Data;
-    uint64_t size;
+    size_t size;
 
 public:
     Matrix () {};
-    Matrix (uint64_t n);
+    Matrix (size_t n);
     ~Matrix() {};
 
-    void Matrix_resize (uint64_t n);
-    void Add_El (uint64_t i, uint64_t j, double data);
+    void Matrix_resize (size_t n);
+    void Add_El (size_t i, size_t j, double data);
     std::vector<double> Matrix_multiply_vector (const std::vector<double> &v) const;
     void print() const;
     void print_data() const;
